fix: addressing reviewer comments
