+++ config/experiment/ppo_atari_pz/observe_pong_pz.yaml
@@ -7,25 +7,13 @@ defaults:
   - override /services/experiment_tracker@run.experiment_tracker: simple
 run:
   players:
-    # - name: first_0
-    #   implementation: actors.ppo_atari_pz.PPOActor
-    #   agent_config:
-    #     model_id: compassionate_faraday_0_model
-    #     model_iteration: -1
-    # - name: second_0
-    #   implementation: actors.ppo_atari_pz.PPOActor
-    #   agent_config:
-    #     model_id: compassionate_faraday_0_model
-    #     model_iteration: -1
-
     - name: first_0
       implementation: actors.ppo_atari_pz.PPOActor
       agent_config:
-        model_id: nostalgic_pare_0_model
+        model_id: ???
         model_iteration: -1
     - name: second_0
       implementation: actors.ppo_atari_pz.PPOActor
       agent_config:
-        model_id: nostalgic_pare_0_model
+        model_id: ???
         model_iteration: -1
-
